feat(SBOMER-298): Correct AtlasClient exception mapping
Create new AtlasClientExceptionMapper as exception mapping was broken.
